feat: list creation to db

+++ backend/controllers/board.controller.ts
@@ -1,10 +1,12 @@
 import type { Request, Response } from "express";
 import Board from "../models/board.model.js";
+import List from "../models/list.model.js";
 
 export class BoardController {
-  static async getAllBoards(req: Request, res: Response) {
+  static async getAllBoardsByUser(req: Request, res: Response) {
     try {
-      const boards = await Board.find();
+      const userId = req.params.userId;
+      const boards = await Board.find({ userId }).sort({ createdAt: -1 });
       res.json(boards);
     } catch (err: unknown) {
       const error = err as Error;
@@ -14,8 +16,8 @@ export class BoardController {
 
   static async createBoard(req: Request, res: Response) {
     try {
-      const { name } = req.body;
-      const newBoard = new Board({ name });
+      const { name, userId } = req.body;
+      const newBoard = new Board({ name, userId });
       await newBoard.save();
       res.status(201).json(newBoard);
     } catch (err: unknown) {
@@ -26,21 +28,38 @@ export class BoardController {
 
   static async getBoardById(req: Request, res: Response) {
     try {
-      const board = await Board.findById(req.params.id);
+      const board = await Board.findById(req.params.boardId);
       if (!board) return res.status(404).json({ error: "Board not found" });
-      res.json(board);
+      if (board.userId.toString() !== req.params.userId)
+        return res.status(403).json({ error: "Unauthorized" });
+
+      const populatedBoard = await Board.findById(req.params.boardId).populate({
+        path: "lists",
+        populate: { path: "cards" }
+      });
+      ;
+      if (!populatedBoard) return res.status(404).json({ error: "Board not found" });
+      res.json(populatedBoard);
     } catch (err: unknown) {
       const error = err as Error;
       res.status(500).json({ error: error.message });
     }
   }
 
-  static async updateBoardLists(req: Request, res: Response) {
+  static async addListToBoard(req: Request, res: Response) {
     try {
-      const { title, lists } = req.body;
+      const board = await Board.findById(req.params.id);
+      if (!board) {
+        return res.status(404).json({ message: "Board not found" });
+      }
+      if (board.userId.toString() !== req.body.userId) {
+        return res.status(403).json({ message: "Unauthorized" });
+      }
+      const newList = await new List({ name: req.body.name, boardId: req.params.id }).save();
+
       const updatedBoard = await Board.findByIdAndUpdate(
         req.params.id,
-        { title, lists },
+        { $push: { lists: newList._id } },
         { new: true }
       );
       if (!updatedBoard)
@@ -54,7 +73,11 @@ export class BoardController {
 
   static async updateBoardName(req: Request, res: Response) {
     try {
-      const { name } = req.body;
+      const { name, userId } = req.body;
+      const boardToUpdate = await Board.findById(req.params.id)
+      if (boardToUpdate?.userId.toString() !== userId)
+        return res.status(403).json({ error: "Unauthorized" });
+
       const updatedBoard = await Board.findByIdAndUpdate(
         req.params.id,
         { name },
@@ -71,6 +94,12 @@ export class BoardController {
 
   static async deleteBoard(req: Request, res: Response) {
     try {
+      const boardToDelete = await Board.findById(req.params.id)
+      if (!boardToDelete)
+        return res.status(404).json({ error: "Board not found" });
+      if (boardToDelete.userId.toString() !== req.params.userId)
+        return res.status(403).json({ error: "Unauthorized" });
+
       const deletedBoard = await Board.findByIdAndDelete(req.params.id);
       if (!deletedBoard)
         return res.status(404).json({ error: "Board not found" });
